+++ structurizr-dsl/src/main/java/com/structurizr/dsl/DeploymentNodeParser.java
@@ -1,6 +1,11 @@
 package com.structurizr.dsl;
 
 import com.structurizr.model.DeploymentNode;
+import com.structurizr.model.Model;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 final class DeploymentNodeParser extends AbstractParser {
 
@@ -13,55 +18,14 @@ final class DeploymentNodeParser extends AbstractParser {
     private static final int INSTANCES_INDEX = 5;
 
     DeploymentNode parse(DeploymentEnvironmentDslContext context, Tokens tokens, Archetype archetype) {
-        // deploymentNode <name> [description] [technology] [tags] [instances]
-
-        if (tokens.hasMoreThan(INSTANCES_INDEX)) {
-            throw new RuntimeException("Too many tokens, expected: " + GRAMMAR);
-        }
-
-        if (!tokens.includes(NAME_INDEX)) {
-            throw new RuntimeException("Expected: " + GRAMMAR);
-        }
-
-        DeploymentNode deploymentNode = null;
-        String name = tokens.get(NAME_INDEX);
-
-        String description = archetype.getDescription();
-        if (tokens.includes(DESCRIPTION_INDEX)) {
-            description = tokens.get(DESCRIPTION_INDEX);
-        }
-
-        String technology = archetype.getTechnology();
-        if (tokens.includes(TECHNOLOGY_INDEX)) {
-            technology = tokens.get(TECHNOLOGY_INDEX);
-        }
-
-        deploymentNode = context.getWorkspace().getModel().addDeploymentNode(context.getEnvironment(), name, description, technology);
-
-        String[] tags = archetype.getTags().toArray(new String[0]);
-        if (tokens.includes(TAGS_INDEX)) {
-            tags = tokens.get(TAGS_INDEX).split(",");
-        }
-        deploymentNode.addTags(tags);
-
-        deploymentNode.addProperties(archetype.getProperties());
-        deploymentNode.addPerspectives(archetype.getPerspectives());
-
-        String instances = "1";
-        if (tokens.includes(INSTANCES_INDEX)) {
-            instances = tokens.get(INSTANCES_INDEX);
-            deploymentNode.setInstances(instances);
-        }
-
-        if (context.hasGroup()) {
-            deploymentNode.setGroup(context.getGroup().getName());
-            context.getGroup().addElement(deploymentNode);
-        }
-
-        return deploymentNode;
+        return parse(context, null, tokens, archetype);
     }
 
     DeploymentNode parse(DeploymentNodeDslContext context, Tokens tokens, Archetype archetype) {
+        return parse(null, context, tokens, archetype);
+    }
+
+    DeploymentNode parse(DeploymentEnvironmentDslContext deploymentEnvironmentDslContext, DeploymentNodeDslContext deploymentNodeDslContext, Tokens tokens, Archetype archetype) {
         // deploymentNode <name> [description] [technology] [tags] [instances]
 
         if (tokens.hasMoreThan(INSTANCES_INDEX)) {
@@ -85,14 +49,28 @@ final class DeploymentNodeParser extends AbstractParser {
             technology = tokens.get(TECHNOLOGY_INDEX);
         }
 
-        DeploymentNode parent = context.getDeploymentNode();
-        deploymentNode = parent.addDeploymentNode(name, description, technology);
+        if (deploymentEnvironmentDslContext != null) {
+            // add a root deployment node
+            deploymentNode = deploymentEnvironmentDslContext.getWorkspace().getModel().addDeploymentNode(deploymentEnvironmentDslContext.getEnvironment(), name, description, technology);
 
-        String[] tags = archetype.getTags().toArray(new String[0]);
-        if (tokens.includes(TAGS_INDEX)) {
-            tags = tokens.get(TAGS_INDEX).split(",");
+            if (deploymentEnvironmentDslContext.hasGroup()) {
+                deploymentNode.setGroup(deploymentEnvironmentDslContext.getGroup().getName());
+                deploymentEnvironmentDslContext.getGroup().addElement(deploymentNode);
+            }
+        } else {
+            deploymentNode = deploymentNodeDslContext.getDeploymentNode().addDeploymentNode(name, description, technology);
+
+            if (deploymentNodeDslContext.hasGroup()) {
+                deploymentNode.setGroup(deploymentNodeDslContext.getGroup().getName());
+                deploymentNodeDslContext.getGroup().addElement(deploymentNode);
+            }
         }
-        deploymentNode.addTags(tags);
+
+        List<String> tags = new ArrayList<>(archetype.getTags());
+        if (tokens.includes(TAGS_INDEX)) {
+            tags.addAll(Arrays.asList(tokens.get(TAGS_INDEX).split(",")));
+        }
+        deploymentNode.addTags(tags.toArray(new String[0]));
 
         deploymentNode.addProperties(archetype.getProperties());
         deploymentNode.addPerspectives(archetype.getPerspectives());
@@ -101,11 +79,6 @@ final class DeploymentNodeParser extends AbstractParser {
         if (tokens.includes(INSTANCES_INDEX)) {
             instances = tokens.get(INSTANCES_INDEX);
             deploymentNode.setInstances(instances);
-        }
-
-        if (context.hasGroup()) {
-            deploymentNode.setGroup(context.getGroup().getName());
-            context.getGroup().addElement(deploymentNode);
         }
 
         return deploymentNode;