+++ structurizr-dsl/src/test/java/com/structurizr/dsl/PersonParserTests.java
@@ -63,4 +63,19 @@ class PersonParserTests extends AbstractTests {
         assertEquals("Element,Person,Tag 1,Tag 2", user.getTags());
     }
 
+    @Test
+    void test_parse_CreatesAPersonWithADescriptionAndTagsBasedUponAnArchetype() {
+        archetype = new Archetype("name", "type");
+        archetype.setDescription("Default Description");
+        archetype.addTags("Default Tag");
+
+        parser.parse(context(), tokens("person", "User", "Description", "Tag 1, Tag 2"), archetype);
+
+        assertEquals(1, model.getElements().size());
+        Person user = model.getPersonWithName("User");
+        assertNotNull(user);
+        assertEquals("Description", user.getDescription()); // overridden from archetype
+        assertEquals("Element,Person,Default Tag,Tag 1,Tag 2", user.getTags());
+    }
+
 }